No issue: increase backfill iteration count to 30.
Around 6/9/22, the variance of the results increased from ~20ms to ~50ms for all
tests but Focus View, which was always noisy. To reduce the noise to help catch
regressions, we increase the iteration count to 30.

Note: I've been uploading results on 30 iterations for several months from local
runs: it's just taken me a long time to commit this change.